Allow serialization of all types that implement Marshal
Previously, this was limited to product types that contain certain types
like Int32, Word32, ...

This change has some consequences: newtypes of the previously types are
now supported, as well as other product types inside other product types.
